refactor: move github.ts in frameworks.ts and templates.ts
